requirements.txt
flask==3.1.0
joblib==1.4.2
numpy==1.25.2
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
scikit_learn==1.5.2
